Fix git index entry for Buddycloud.json
Somehow a recent update (probably made from a Windows machine with a
filesystem that doesn't properly distinguish between files with
different cases) merged in a change that added buddycloud.json but left
this file lying around.  This leads to weird behaviors in fresh git
clones on filesystems that have case-sensitivity.  Removing the
duplicate file; Buddycloud.json appears to have less data so that is the
one I'm removing.