mark Project Byzantium as dead
As one of the original founders of Project Byzantium, I can tell you with certainty that the project is long dead. It has not been maintained in over 5 years. None of the core developers have any interest in reviving it. There are known bugs and security vulnerabilities which will not be fixed.